+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.4.5 (25 May 2026)
+
+### Changed
+
+- only allow a single instance of `kiss-tnc-bridge` to run, running multiple
+  instances concurrently, especially fron the same config file can cause
+  race conditions and malformed data sent to the TCP TNC.
+
+
 ## 0.4.4 (25 May 2026)
 
 ### Improved